tests/wpt/tests/css/css-color-adjust/animation/color-scheme-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<link rel=help href="https://drafts.csswg.org/css-color-adjust/#color-scheme-prop">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'color-scheme',
  from: 'initial',
  to: 'dark'
});
</script>
